requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
six==1.16.0
Werkzeug==3.0.1
wsproto==1.2.0
